Add Datasource\LocatorInface.
ORM\LocatorInterface now extends the above interface.
